requirements.txt
fastapi
uvicorn
python-dotenv
gitpython
neo4j
pinecone
tree-sitter
tree-sitter-language-pack
google-generativeai
